Move getSocialIconForWidget to widget-icon
